Improve structure of stats
Two main things to hopefully improve legibility. The first is the
convert content_length into an actual struct so that we are dealing with
a type of ContentLength. The second is to implement a trait for
duration to help convert duration to milliseconds. This is used in both
display and calculating some values.