Refactor to improve encapsulation and readability
Refactored several things in order to try to improve readability.
Starting with the ContentLength, I moved that to a separate module and
implemented the Add trait so that the facts and summary can opearate on
a nominative type. From there I factored out the summarizing of
durations into it's own struct that contains the sorting of the
durations along with analyzing it's statistics. This reduced the
complexity of the Summary struct while encapsulating the logic of
determining the stats.