Add engine support for hyper
In order to compare reqwest to hyper I wanted to have both inside the
same compiled binary. I also wanted to be able to specify which engine
to use so I can run it twice. To do this I did the following:

1. Changed Facts to no longer be dependent on the reqwest types
2. Implemented a hyper enginer function
3. Set up an engine argument for specifying which engine to use.

In the end, this pulls in a few more dependencies but provided a clear
methodology for running either client.