Refactored out the collector and runner
To make the code more testable and to isolate some of the features,
I've extracted the collector and runner into their own modules. I've
also moved the message into it's own and created the idea of a "plan"

This has had the effect of minimizing the main code but also allowing
for future generics for each of the types. One plan for the future is
replace the engine with a trait so that there can be multiple
implementations of the engine. This will work even better with the
runner if we make the engine generic over the "facts" it produces.